newDashboard -> dashboard for router.push
