--- src/app/newDashboard/page.jsx
@@ -1,11 +1,0 @@
-'use client';
-
-const NewDashboardPage = () => {
-    return (
-        <div className="d-flex align-items-center justify-content-center h-100">
-            <h2>Migrating SPA Design to New Dashboard</h2>
-        </div>
-    );
-};
-
-export default NewDashboardPage;
